refactor(clicommandfactory): migrate CLI commands to Factory Pattern architecture
- Migrate Ollama and SageMaker commands from Command Module to Factory Pattern
- Create OllamaCommandFactory with 8 subcommands (list-models, pull, remove, etc.)
- Create SageMakerCommandFactory with 9 subcommands (status, test, setup, etc.)
- Update CLICommandFactory with delegation methods for both factories
- Update main CLI entry point to use factory-created commands
- Remove legacy command module files (src/cli/commands/ollama.ts, sagemaker.ts)
- Add comprehensive CLI Command Factory Developer Guide

Benefits:
- Improved code organization and maintainability
- Enhanced type safety with proper TypeScript interfaces
- Consistent error handling and user experience patterns
- Built-in security patterns for credential management
- Scalable architecture for future CLI command additions
